+++ CHANGELOG.rst
@@ -13,14 +13,15 @@ v3.0.0
 
    | It only needs PIL library to compare images and generate the differences images
    | Old PerceptualDiff and Magick engines have been removed
-   | Config property 'visualtests_engine' in [Server] section has been removed
+   | Config property `engine` in [VisualTests] section has been removed
    | Images distance calculation method has changed, it is recommended to review thresholds in tests
 
 - Now `gecko_driver_path`, `chrome_driver_path`, `explorer_driver_path` and `edge_driver_path` config properties
   in [Driver] section are optional, due to new SeleniumManager feature, that downloads drivers automatically
 - New optional config property `safari_driver_path` in [Driver] section to configure Safari driver
-- New optional config property `options` in [Chromer] section to configure Chrome options instead of using old
+- New optional config property `options` in [Chrome] section to configure Chrome options instead of using old
   property `goog:chromeOptions` in [Capabilities] section.
+- New optional config property `base_path` in [Server] section to allow using old Selenium 3 or Appium 1 remote servers
 - Update [RANDOM_PHONE_NUMBER] replacement using new `DataGenerator` class
 - Remove support for lettuce tests
 - Remove deprecated parameter `context` from `map_param` and POEditor methods